//Vérification des dates et début pour historiser les questions supprimées.

+++ src/View/question/list.php
@@ -2,6 +2,8 @@
 <div class="container_liste_question">
     <div class="container_question">
 <?php
+
+
 echo "<div class='container_question1'>";
     echo "<div class='titrelistequestion'>";
         echo "<div id='apparait' class='cacher'>";
@@ -13,17 +15,19 @@ echo "<div class='container_question1'>";
 echo "</div>";
 
 foreach ($questions as $question) {
-    $questNonFormater = $question->getIdQuestion();
-    $questFormater = rawurlencode($questNonFormater);
+    if ($question->isActif() == false) {
+        $questNonFormater = $question->getIdQuestion();
+        $questFormater = rawurlencode($questNonFormater);
 
-    echo "<div class='questions'>";
-    echo "<a id='titrequestion' href=\"frontController.php?controller=question&action=read&id_question={$questFormater}\"> Question {$question->getIdQuestion()} :\n".  htmlspecialchars ( $question->getIntitule() ) . " </a>";
-    echo "<div class='question_update'>";
-    echo "<a href=\"frontController.php?controller=question&action=update&id_question={$questFormater}\"> <i class='fa-solid fa-pencil'></i> </a>";
-    echo "<a id=\"confirmation\" onclick=\"return confirmation()\" href=\"frontController.php?controller=question&action=delete&id_question={$questFormater}\"> <i class='fa-solid fa-trash'></i></a>";
-    echo "</div>";
-    echo "</div>";
+        echo "<div class='questions'>";
+        echo "<a id='titrequestion' href=\"frontController.php?controller=question&action=read&id_question={$questFormater}\"> Question {$question->getIdQuestion()} :\n" . htmlspecialchars($question->getIntitule()) . " </a>";
+        echo "<div class='question_update'>";
+        echo "<a href=\"frontController.php?controller=question&action=update&id_question={$questFormater}\"> <i class='fa-solid fa-pencil'></i> </a>";
+        echo "<a id=\"confirmation\" onclick=\"return confirmation()\" href=\"frontController.php?controller=question&action=delete&id_question={$questFormater}\"> <i class='fa-solid fa-trash'></i></a>";
+        echo "</div>";
+        echo "</div>";
 //onclick=\"validation()\"
+    }
 }
 ?>
     </div>
